+++ ax3l/app/snakelab/prompts/FirstContact.py
@@ -6,7 +6,10 @@ class FirstContact(Prompt):
 
         MSG = (
             "This experiment is an implementation of Patrick Loeber's Famous "
-            "AI Snake Game. You are in charge of the experiment."
+            "AI Snake Game. You are responsible choosing values that will achieve "
+            "the best possible high score. Sometimes your available choices all "
+            "look bad. You can still gather important data by running simulations "
+            "in these ranges. You may also discover a new higher plateau!"
         )
 
         super().__init__(MSG)